#7- BOJ 2588 문제 세팅
